requirements.txt
gradio>=4.0.0
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
